requirements.txt
torch>=2.4.1
torchvision>=0.19.0
numpy>=1.24.0
matplotlib>=3.7.0
pandas>=2.0.0
scipy>=1.10.0
tqdm>=4.65.0
tensorboard>=2.14.0
pillow>=10.0.0
jupyter>=1.0.0
ipykernel>=6.25.2
opencv-python>=4.8.1
diffusers>=0.28.0
accelerate>=0.29.0
einops>=0.7.0